pyproject.toml
[tool.poetry]
name = "FileManager"
version = "1.0.0"
description = "FileManager"
authors = []

[tool.poetry.dependencies]
python = "^3.13"
django = "5.2.7"
psycopg2-binary = "2.9.10"
django-filter = "^24.3"
djangorestframework = "^3.16.1"
djangorestframework-stubs = "^3.15.2"  # typing
python-dotenv = "^1.0.1"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
django-extensions = "^3.2.3"  # shell_plus
django-cors-headers = "^4.6.0"
django-storages = { version = "^1.14.6", extras = ["boto3"] }
boto3 = "^1.40.55"
drf-spectacular-sidecar = "^2025.10.1"
drf-spectacular = "^0.28.0"
filetype = "^1.2.0"
pillow = "^12.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"
flake8-bugbear = "^24.4.26"
pre-commit = "^4.0.1"

[tool.black]
line-length = 120
# Needed as of black 23.12.1 to properly split long stings
preview = true

[tool.isort]
profile = "black"
line_length = 120
skip = ['.venv', '.pytest_cache']

[tool.flake8]
max-line-length = 120
extend-select = ['B950']
# E203, E501, E701 are rules incompatiblie with black and are exclued as recommended by black documentation
extend-ignore = ['E501', 'E701', 'E203']
extend-exclude = ['.venv', '.pytest_cache']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "conf.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
#addopts = "--create-db --reuse-db"
addopts = "--reuse-db"